MedikeaAIBackend/requirements.txt
Django
django-jazzmin
djangorestframework
djangorestframework-simplejwt
numpy
pandas
PyJWT
PyMySQL
python-dateutil
python-decouple
python-dotenv
APScheduler
psycopg
terminado
text-unidecode
textblob                    
transformers 
urllib3
websocket-client 
tensorflow
keras
scikit-learn
drf-spectacular
gensim
xformers
google-generativeai
terminado
text-unidecode
textblob
tf_keras2                     
transformers 
urllib3
websocket-client 
tensflow

